# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OrcaSlicer API",
    "version": "1.1.0",
    "description": "REST API for slicing 3D models (STL, OBJ, AMF, 3MF and STEP/IGES/BREP CAD files) with OrcaSlicer. Manage profiles, inspect geometry, convert models and read print statistics. The same engine is exposed as the 'osa' terminal CLI \u2014 see CLI.md."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/profiles/{category}": {
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Upload a profile (printer, preset, or filament)",
        "description": "Upload a new printer, preset, or filament profile (OrcaSlicer JSON format).",
        "parameters": [
          {
            "$ref": "#/components/parameters/Category"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "name"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "OrcaSlicer profile JSON file"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name for the profile (alphanumeric)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "List all profiles in a category",
        "description": "List all uploaded profiles for a given category (printers, presets, filaments).",
        "parameters": [
          {
            "$ref": "#/components/parameters/Category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of profile names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{category}/{name}": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get a specific profile by name",
        "description": "Get a specific printer, preset, or filament profile by name.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Category"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Profile name (alphanumeric)"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "summary": "Delete a specific profile by name",
        "description": "Deletes a specific printer, preset, or filament profile by name.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Category"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Profile name (alphanumeric)"
          }
        ],
        "responses": {
          "204": {
            "description": "Profile deleted"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profiles"
        ],
        "summary": "Patch settings on a stored profile",
        "description": "Applies key=value setting changes to a stored profile, the API equivalent of the CLI 'config set'.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Category"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Profile name (alphanumeric)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "set"
                ],
                "properties": {
                  "set": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "key=value assignments, e.g. layer_height=0.28"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      }
    },
...
    "/tools/info": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Inspect a model's geometry",
        "description": "Returns the bounding box, triangle count, part count, manifold status and volume. CAD files are triangulated first.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Model file (STL, OBJ, AMF, 3MF, STEP, IGES or BREP)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model geometry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid file"
          }
        }
      }
    },
    "/tools/convert": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Convert a model to STL or 3MF",
        "description": "Converts an uploaded model and streams the result. STEP/IGES/BREP CAD files are triangulated with OpenCASCADE.",
        "parameters": [
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "stl",
                "3mf"
              ],
              "default": "stl"
            },
            "description": "Target format"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Model file (STL, OBJ, AMF, 3MF, STEP, IGES or BREP)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Converted file",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid file / format"
          }
        }
      }
    },
    "/tools/inspect": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Read print statistics from a sliced file",
        "description": "Extracts print time, filament usage, layer count and pricing metrics from a G-code or 3MF file.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Sliced G-code (.gcode) or 3MF (.3mf) file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Print statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceMetadata"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid file"
          }
        }
      }
    },
...
    "/presets/{vendor}/{kind}": {
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "List a vendor's system presets",
        "parameters": [
          {
            "name": "vendor",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "machine",
                "process",
                "filament"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preset names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid kind"
          }
        }
      }
    },
    "/presets/{vendor}/{kind}/{name}": {
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "Get a flattened system preset",
        "description": "Returns the preset with its inherits chain fully resolved.",
        "parameters": [
          {
            "name": "vendor",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "machine",
                "process",
                "filament"
              ]
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flattened preset JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Preset not found"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Category": {
        "name": "category",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "printers",
            "presets",
            "filaments"
          ]
        },
        "description": "Profile category (printers, presets, filaments)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "SliceMetadata": {
        "type": "object",
        "properties": {
          "printTime": {
            "type": "number",
            "description": "Estimated print time in seconds"
          },
          "filamentUsedG": {
            "type": "number",
            "description": "Filament used in grams"
          },
          "filamentUsedMm": {
            "type": "number",
            "description": "Filament used in millimeters"
          }
        }
      },
...
      "ModelInfo": {
        "type": "object",
        "properties": {
          "sizeX": {
            "type": "number"
          },
          "sizeY": {
            "type": "number"
          },
          "sizeZ": {
            "type": "number"
          },
          "minX": {
            "type": "number"
          },
          "minY": {
            "type": "number"
          },
          "minZ": {
            "type": "number"
          },
          "maxX": {
            "type": "number"
          },
          "maxY": {
            "type": "number"
          },
          "maxZ": {
            "type": "number"
          },
          "facets": {
            "type": "integer",
            "description": "Triangle count"
          },
          "parts": {
            "type": "integer",
            "description": "Disconnected part count"
          },
          "manifold": {
            "type": "boolean"
          },
          "volume": {
            "type": "number",
            "description": "Volume in mm3"
          }
        }
      }
    }
  }
}